:art: move model definitions for `connections`
